🔧 Improve typing (#101)
## Overview

I naively implemented typing. It was time to give it some much needed attention. Now IDEs and mypy will correctly interpret types for all methods on Commands and CommandsAsync except for query_multiple and query_multiple_async. I have a TODO to explore variadic typevars (TypeVarTuple) for this in the future.

## Changes

Leverage TypeVar to provide a better typing experience on Commands and CommandsAsync
Add an annotations test to CI for validating some of the more complex type annotations are outputting what I would expect